# api/openapi.yaml
openapi: 3.0.3
info:
  title: ReqViz API
  version: 0.1.0
  description: Backend API for ReqViz — Visual OpenAPI Explorer

servers:
  - url: http://localhost:3000/api
    description: Local Development

paths:
  /specs:
    post:
      operationId: uploadSpec
      summary: Upload an OpenAPI specification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              description: Raw OpenAPI spec as JSON
      responses:
        '201':
          description: Spec uploaded and parsed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpecSummary'
        '400':
          $ref: '#/components/responses/ValidationError'

  /specs/{id}:
    get:
      operationId: getSpec
      summary: Get a stored specification
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Full spec with parsed metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParsedSpec'
        '404':
          $ref: '#/components/responses/NotFound'

  /proxy:
    post:
      operationId: proxyRequest
      summary: CORS proxy for API requests
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProxyRequest'
      responses:
        '200':
          description: Proxied response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProxyResponse'

  /demos:
    get:
      operationId: listDemos
      summary: List available demo API specs
      responses:
        '200':
          description: Array of demo metadata
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DemoInfo'

  /demos/{slug}:
    get:
      operationId: getDemoSpec
      summary: Get the raw OpenAPI spec for a demo
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Raw demo spec as JSON
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '404':
          $ref: '#/components/responses/NotFound'

  /demos/{slug}/load:
    post:
      operationId: loadDemo
      summary: Parse and store a demo spec, returning its summary
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Demo spec parsed and stored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpecSummary'
        '404':
          $ref: '#/components/responses/NotFound'

  /health:
    get:
      operationId: healthCheck
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [ok]

components:
  schemas:
    SpecSummary:
      type: object
      required: [id, title, version, endpointCount, schemaCount]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        version:
          type: string
        endpointCount:
          type: integer
        schemaCount:
          type: integer
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time

    ParsedSpec:
      type: object
      required: [id, raw, summary]
      properties:
        id:
          type: string
          format: uuid
        raw:
          type: object
          additionalProperties: true
          description: Original OpenAPI spec as JSON
        summary:
          $ref: '#/components/schemas/SpecSummary'

    ProxyRequest:
      type: object
      required: [method, url]
      properties:
        method:
          type: string
          enum: [GET, POST, PUT, PATCH, DELETE, HEAD, OPTIONS]
        url:
          type: string
          format: uri
        headers:
          type: object
          additionalProperties:
            type: string
        body:
          description: Request body (any type)

    ProxyResponse:
      type: object
      required: [status, headers]
      properties:
        status:
          type: integer
        headers:
          type: object
          additionalProperties:
            type: string
        body:
          description: Response body (any type)
        durationMs:
          type: integer

    DemoInfo:
      type: object
      required: [slug, title, description]
      properties:
        slug:
          type: string
        title:
          type: string
        description:
          type: string

    ValidationError:
      type: object
      properties:
        error:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string

  responses:
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string